+++ include/databento/live_blocking.hpp
@@ -45,8 +45,8 @@ class LiveBlocking {
     return heartbeat_interval_;
   }
   databento::Compression Compression() const { return compression_; }
-  std::optional<databento::SlowReadBehavior> SlowReadBehavior() const {
-    return slow_read_behavior_;
+  std::optional<databento::SlowReaderBehavior> SlowReaderBehavior() const {
+    return slow_reader_behavior_;
   }
   const std::vector<LiveSubscription>& Subscriptions() const { return subscriptions_; }
   std::vector<LiveSubscription>& Subscriptions() { return subscriptions_; }
@@ -99,14 +99,14 @@ class LiveBlocking {
                std::optional<std::chrono::seconds> heartbeat_interval,
                std::size_t buffer_size, std::string user_agent_ext,
                databento::Compression compression,
-               std::optional<databento::SlowReadBehavior> slow_read_behavior);
+               std::optional<databento::SlowReaderBehavior> slow_reader_behavior);
   LiveBlocking(ILogReceiver* log_receiver, std::string key, std::string dataset,
                std::string gateway, std::uint16_t port, bool send_ts_out,
                VersionUpgradePolicy upgrade_policy,
                std::optional<std::chrono::seconds> heartbeat_interval,
                std::size_t buffer_size, std::string user_agent_ext,
                databento::Compression compression,
-               std::optional<databento::SlowReadBehavior> slow_read_behavior);
+               std::optional<databento::SlowReaderBehavior> slow_reader_behavior);
 
   std::string DetermineGateway() const;
   std::uint64_t Authenticate();
@@ -133,7 +133,7 @@ class LiveBlocking {
   const VersionUpgradePolicy upgrade_policy_;
   const std::optional<std::chrono::seconds> heartbeat_interval_;
   const databento::Compression compression_;
-  const std::optional<databento::SlowReadBehavior> slow_read_behavior_;
+  const std::optional<databento::SlowReaderBehavior> slow_reader_behavior_;
   detail::LiveConnection connection_;
   std::uint32_t sub_counter_{};
   std::vector<LiveSubscription> subscriptions_;