+++ tests/src/live_blocking_tests.cpp
@@ -56,17 +56,17 @@ TEST_F(LiveBlockingTests, TestAuthentication) {
                                   .BuildBlocking();
 }
 
-TEST_F(LiveBlockingTests, TestAuthenticationWithSlowReadBehavior) {
+TEST_F(LiveBlockingTests, TestAuthenticationWithSlowReaderBehavior) {
   constexpr auto kTsOut = false;
-  constexpr auto kSlowReadBehavior = SlowReadBehavior::Warn;
-  const mock::MockLsgServer mock_server{dataset::kXnasItch, kTsOut, kSlowReadBehavior,
+  constexpr auto kSlowReaderBehavior = SlowReaderBehavior::Warn;
+  const mock::MockLsgServer mock_server{dataset::kXnasItch, kTsOut, kSlowReaderBehavior,
                                         [](mock::MockLsgServer& self) {
                                           self.Accept();
                                           self.Authenticate();
                                         }};
 
   const LiveBlocking target = builder_.SetDataset(dataset::kXnasItch)
-                                  .SetSlowReadBehavior(kSlowReadBehavior)
+                                  .SetSlowReaderBehavior(kSlowReaderBehavior)
                                   .SetAddress(kLocalhost, mock_server.Port())
                                   .BuildBlocking();
 }