+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## 0.47.1 - TBD
 
 ### Enhancements
-- Added `SlowReadBehavior` enum and `LiveBuilder::SetSlowReadBehavior()` to configure
+- Added `SlowReaderBehavior` enum and `LiveBuilder::SetSlowReaderBehavior()` to configure
   gateway behavior when client falls behind
-- Added `SlowReadBehavior()` getter to `LiveBlocking` and `LiveThreaded`
+- Added `SlowReaderBehavior()` getter to `LiveBlocking` and `LiveThreaded`
 
 ### Bug fixes
 - Added conversion for missing schemas for function `RTypeFromSchema`